requirements.txt
asgiref==3.7.2
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.2.1
cryptography==41.0.1
decorator==5.1.1
Django==4.2.1
django-environ==0.10.0
django-filter==23.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
fabric==3.1.0
Faker==18.9.0
idna==3.4
inflection==0.5.1
invoke==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
openapi-codec==1.3.2
paramiko==3.2.0
pycparser==2.21
PyJWT==2.7.0
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.2
